moved test app to tests/ directory
