+++ gulpfile.js
@@ -2,7 +2,7 @@ var gulp = require("gulp");
 var babel = require("gulp-babel");
 
 gulp.task("default", function () {
-  return gulp.src("lib/**/*.js")
+  return gulp.src(["src/**/*.js", "src/**/*.jsx"])
     .pipe(babel())
     .pipe(gulp.dest("dist"));
 });
