+++ tests/notes-app/.gitignore
@@ -1,0 +1,4 @@
+terminal.glue
+build/*
+node_modules/*
+npm-debug.log
