+++ tests/notes-app/index.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <div id="app">Loading...</div>
-    <script src="build/bundle.js"></script>
+    <script src="bundle.js"></script>
   </body>
 </html>
