+++ tests/notes-app/webpack.config.js
